requirements.txt
requests
customtkinter
Pillow
py7zr
python-dotenv
pytest
